+++ dev/Layer.js
@@ -13,7 +13,11 @@ class Layer {
 
     assignPrev (layer) {
         this.prevLayer = layer
-        this.neurons.forEach(neuron => neuron.init(layer.size, this.adaptiveLR, this.rho))   
+        this.neurons.forEach(neuron => neuron.init(layer.size, {
+            adaptiveLR: this.adaptiveLR,
+            activationConfig: this.activationConfig,
+            eluAlpha: this.eluAlpha
+        }))
     }
 
     forward (data) {
@@ -22,7 +26,7 @@ class Layer {
 
             neuron.sum = neuron.bias
             this.prevLayer.neurons.forEach((pNeuron, pni) => neuron.sum += pNeuron.activation * neuron.weights[pni])
-            neuron.activation = this.activation(neuron.sum)
+            neuron.activation = this.activation(neuron.sum, false, neuron)
         })
     }
 
@@ -32,7 +36,7 @@ class Layer {
             if(typeof expected !== "undefined") {
                 neuron.error = expected[ni] - neuron.activation
             }else {
-                neuron.derivative = this.activation(neuron.sum, true)
+                neuron.derivative = this.activation(neuron.sum, true, neuron)
                 neuron.error = neuron.derivative * this.nextLayer.neurons.map(n => n.error * n.weights[ni])
                                                                          .reduce((p,c) => p+c, 0)
             }