+++ dev/NetMath.js
@@ -4,10 +4,42 @@ class NetMath {
     
     // Activation functions
     static sigmoid (value, prime) {
-        return prime ? NetMath.sigmoid(value)*(1-NetMath.sigmoid(value))
-                     : 1/(1+Math.exp(-value))
+        const val = 1/(1+Math.exp(-value))
+        return prime ? val*(1-val)
+                     : val
     }
 
+    static tanh (value, prime) {
+        const exp = Math.exp(2*value)
+        return prime ? 4/Math.pow(Math.exp(value)+Math.exp(-value), 2) || 1e-18
+                     : (exp-1)/(exp+1) || 1e-18
+    }
+
+    static relu (value, prime) {
+        return prime ? value > 0 ? 1 : 0
+                     : Math.max(value, 0)
+    }
+
+    static lrelu (value, prime) {
+        return prime ? value > 0 ? 1 : this.lreluSlope
+                     : Math.max(this.lreluSlope*Math.abs(value), value)
+    }
+
+    static rrelu (value, prime, neuron) {
+        return prime ? value > 0 ? 1 : neuron.rreluSlope
+                     : Math.max(neuron.rreluSlope, value)   
+    }
+
+    static lecuntanh (value, prime) {
+        return prime ? 1.15333 * Math.pow(NetMath.sech((2/3) * value), 2)
+                     : 1.7159 * NetMath.tanh((2/3) * value)
+    }
+
+    static elu (value, prime, neuron) {
+        return prime ? value >=0 ? 1 : NetMath.elu(value, false, neuron) + neuron.eluAlpha
+                     : value >=0 ? value : neuron.eluAlpha * (Math.exp(value) - 1)
+    }
+    
     // Cost functions
     static crossEntropy (target, output) {
         return output.map((value, vi) => target[vi] * Math.log(value+1e-15) + ((1-target[vi]) * Math.log((1+1e-15)-value)))
@@ -92,6 +124,10 @@ class NetMath {
     static softmax (values) {
         const total = values.reduce((prev, curr) => prev+curr, 0)
         return values.map(value => value/total)
+    }
+
+    static sech (value) {
+        return (2*Math.exp(-value))/(1+Math.exp(-2*value))
     }
 }