+++ test/test.js
@@ -28,7 +28,7 @@ describe("Network", () => {
             beforeEach(() => net = new Network())
 
             it("Defaults the activation to sigmoid and sets the function from NetMath to net.activation", () => {
-                expect(net.activation).to.equal(NetMath.sigmoid)
+                expect(net.activation.name).to.equal("bound sigmoid")
             })
 
             it("Defaults the learning rate to 0.2", () => {
@@ -115,6 +115,91 @@ describe("Network", () => {
                 expect(net2.rho).to.equal(0.9)
                 expect(net2.learningRate).to.equal(0.01)
             })
+
+            it("Defaults the learningRate to 0.001 if the activation is tanh", () => {
+                const net = new Network({activation: "tanh"})
+                expect(net.learningRate).to.equal(0.001)
+            })
+
+            it("Still allows a user to set a learningRate when activation is tanh", () => {
+                const net = new Network({activation: "tanh", learningRate: 0.01})
+                expect(net.learningRate).to.equal(0.01)
+            })
+
+            it("Defaults the learningRate to 0.01 when activation is relu", () => {
+                const net = new Network({activation: "relu"})
+                expect(net.learningRate).to.equal(0.01)
+            })
+
+            it("Still allows a user to set a learningRate when activation is relu", () => {
+                const net = new Network({activation: "relu", learningRate: 0.001})
+                expect(net.learningRate).to.equal(0.001)
+            })
+
+            it("Defaults lreluSlope to -0.0005 when using lrelu activation", () => {
+                const net = new Network({activation: "lrelu"})
+                expect(net.lreluSlope).to.equal(-0.0005)
+            })
+
+            it("Still allows a user to set a lreluSlope when activation is lrelu", () => {
+                const net = new Network({activation: "lrelu", lreluSlope: -0.001})
+                expect(net.lreluSlope).to.equal(-0.001)
+            })
+
+            it("Defaults the learningRate to 0.01 when activation is lrelu", () => {
+                const net = new Network({activation: "lrelu"})
+                expect(net.learningRate).to.equal(0.01)
+            })
+
+            it("Still allows a user to set a learningRate when activation is lrelu", () => {
+                const net = new Network({activation: "lrelu", learningRate: 0.001})
+                expect(net.learningRate).to.equal(0.001)
+            })
+
+            it("Sets the net.activationConfig to the given activation config value.", () => {
+                const net = new Network({activation: "relu"})
+                expect(net.activationConfig).to.equal("relu")
+            })
+
+            it("Defaults the learningRate to 0.01 when activation is rrelu", () => {
+                const net = new Network({activation: "rrelu"})
+                expect(net.learningRate).to.equal(0.01)
+            })
+
+            it("Still allows a user to set a learningRate when activation is rrelu", () => {
+                const net = new Network({activation: "rrelu", learningRate: 0.001})
+                expect(net.learningRate).to.equal(0.001)
+            })
+
+            it("Defaults the learningRate to 0.001 when activation is lecuntanh", () => {
+                const net = new Network({activation: "lecuntanh"})
+                expect(net.learningRate).to.equal(0.001)
+            })
+
+            it("Still allows a user to set a learningRate when activation is lecuntanh", () => {
+                const net = new Network({activation: "lecuntanh", learningRate: 0.01})
+                expect(net.learningRate).to.equal(0.01)
+            })
+
+            it("Defaults the learningRate to 0.01 when activation is elu", () => {
+                const net = new Network({activation: "elu"})
+                expect(net.learningRate).to.equal(0.01)
+            })
+
+            it("Still allows a user to set a learningRate when activation is elu", () => {
+                const net = new Network({activation: "elu", learningRate: 0.001})
+                expect(net.learningRate).to.equal(0.001)
+            })
+
+            it("Defaults the eluAlpha to 1 when activation is elu", () => {
+                const net = new Network({activation: "elu"})
+                expect(net.eluAlpha).to.equal(1)
+            })
+
+            it("Still allows a user to set a eluAlpha when activation is elu", () => {
+                const net = new Network({activation: "elu", eluAlpha: 2})
+                expect(net.eluAlpha).to.equal(2)
+            })
         })
 
         it("Can create a new Network with no parameters", () => expect(new Network()).instanceof(Network))
@@ -177,7 +262,7 @@ describe("Network", () => {
         it("Calls the joinLayer function with each layer when state is constructed", () => {
             const layer1 = new Layer(1)
             const layer2 = new Layer(2)
-            const netThis = {state: "constructed", layers: [layer1, layer2], joinLayer: net.joinLayer}
+            const netThis = {state: "constructed", layers: [layer1, layer2], joinLayer: net.joinLayer, activation: NetMath.sigmoid}
             net.initLayers.bind(netThis)()
 
             expect(netThis.state).to.equal("initialised")
@@ -187,14 +272,14 @@ describe("Network", () => {
         })
 
         it("Calculates reasonable sizes for layers, when state is defined (with a small net)", () => {
-            const netThis = {state: "defined", definedLayers: [Layer, Layer, Layer, Layer], joinLayer: net.joinLayer}
+            const netThis = {state: "defined", definedLayers: [Layer, Layer, Layer, Layer], joinLayer: net.joinLayer, activation: NetMath.sigmoid}
             net.initLayers.bind(netThis, 3, 2)()
             expect(netThis.layers.map(layer => layer.size)).to.deep.equal([3, 5, 3, 2])
             expect(net.joinLayer.callCount).to.equal(4)
         })
 
         it("Calculates reasonable sizes for layers, when state is defined (with a big net)", () => {
-            const netThis = {state: "defined", definedLayers: [Layer, Layer, Layer, Layer, Layer, Layer], joinLayer: net.joinLayer}
+            const netThis = {state: "defined", definedLayers: [Layer, Layer, Layer, Layer, Layer, Layer], joinLayer: net.joinLayer, activation: NetMath.sigmoid}
             net.initLayers.bind(netThis, 784, 10)()
 
             expect(netThis.state).to.equal("initialised")
@@ -203,7 +288,7 @@ describe("Network", () => {
         })
 
         it("Creates three Layers when state is not-defined. First and last layer sizes respective to input/output, middle is in-between", () => {
-            const netThis = {state: "not-defined", joinLayer: net.joinLayer, layers: []}
+            const netThis = {state: "not-defined", joinLayer: net.joinLayer, layers: [], activation: NetMath.sigmoid}
             net.initLayers.bind(netThis, 3, 2)()
 
             expect(netThis.state).to.equal("initialised")
@@ -212,7 +297,7 @@ describe("Network", () => {
         })
 
         it("Creates three Layers when state is not-defined. (the same, but with big net)", () => {
-            const netThis = {state: "not-defined", joinLayer: net.joinLayer, layers: []}
+            const netThis = {state: "not-defined", joinLayer: net.joinLayer, layers: [], activation: NetMath.sigmoid}
             net.initLayers.bind(netThis, 784, 10)()
             expect(netThis.state).to.equal("initialised")
             expect(net.joinLayer).to.have.been.calledThrice
@@ -222,10 +307,10 @@ describe("Network", () => {
         it("Sets the network's activation function to the layers", () => {
             const layer1 = new Layer(2)
             const layer2 = new Layer(3)
-            const net = new Network({layers: [layer1, layer2]})
+            const net = new Network({layers: [layer1, layer2], activation: "sigmoid"})
 
-            expect(layer1.activation).to.equal(NetMath.sigmoid)
-            expect(layer2.activation).to.equal(NetMath.sigmoid)
+            expect(layer1.activation.name).to.equal("bound sigmoid")
+            expect(layer2.activation.name).to.equal("bound sigmoid")
         })
     })
 
@@ -288,6 +373,27 @@ describe("Network", () => {
             net.rho = undefined
             net.joinLayer(layer1)
             expect(layer1.rho).to.be.undefined
+        })
+
+        it("Sets the layer's activationConfig to the net.activationConfig", () => {
+            net.layers = [layer1]
+            net.activationConfig = "test"
+            net.joinLayer(layer1)
+            expect(layer1.activationConfig).to.equal("test")
+        })
+
+        it("Assigns the network's eluAlpha value to the layer, if it exists", () => {
+            net.layers = [layer1]
+            net.eluAlpha = "test"
+            net.joinLayer(layer1)
+            expect(layer1.eluAlpha).to.equal("test")
+        })
+
+        it("Does not set the network's eluAlpha value to the layer if it does not exist", () => {
+            net.layers = [layer1]
+            net.eluAlpha = undefined
+            net.joinLayer(layer1)
+            expect(layer1.eluAlpha).to.be.undefined
         })
     })
 
@@ -812,11 +918,21 @@ describe("Layer", () => {
             expect(layer2.neurons[1].init).to.have.been.calledWith(2)
         })
 
-        it("Calls the neuron's init function with this layer's adaptiveLR", () => {
-            layer2.adaptiveLR = "testStuff"
+        it("Calls the neuron's init function with this layer's adaptiveLR and activationConfig", () => {
+            layer2.adaptiveLR = "test"
+            layer2.activationConfig = "stuff"
             layer2.assignPrev(layer1)
-            expect(layer2.neurons[0].init).to.have.been.calledWith(2, "testStuff")
-            expect(layer2.neurons[1].init).to.have.been.calledWith(2, "testStuff")
+            expect(layer2.neurons[0].init).to.have.been.calledWith(2, sinon.match({"adaptiveLR": "test"}))
+            expect(layer2.neurons[0].init).to.have.been.calledWith(2, sinon.match({"activationConfig": "stuff"}))
+            expect(layer2.neurons[1].init).to.have.been.calledWith(2, sinon.match({"adaptiveLR": "test"}))
+            expect(layer2.neurons[1].init).to.have.been.calledWith(2, sinon.match({"activationConfig": "stuff"}))
+        })
+
+        it("Calls the neuron's init function with this layer's eluAlpha", () => {
+            layer2.eluAlpha = 1
+            layer2.assignPrev(layer1)
+            expect(layer2.neurons[0].init).to.have.been.calledWith(2, sinon.match({"eluAlpha": 1}))
+            expect(layer2.neurons[1].init).to.have.been.calledWith(2, sinon.match({"eluAlpha": 1}))
         })
     })
 
@@ -1000,104 +1116,116 @@ describe("Neuron", () => {
         })
 
         it("Creates a weightGains array if the adaptiveLR parameter is gain, with same size as weights, with 1 values", () => {
-            neuron.init(3, "gain")
+            neuron.init(3, {adaptiveLR: "gain"})
             expect(neuron.weightGains).to.not.be.undefined
             expect(neuron.weightGains).to.have.lengthOf(3)
             expect(neuron.weightGains).to.deep.equal([1,1,1])
         })
 
         it("Creates a biasGain value of 1 if the adaptiveLR parameter is gain", () => {
-            neuron.init(3, "gain")
+            neuron.init(3, {adaptiveLR: "gain"})
             expect(neuron.biasGain).to.equal(1)
         })
 
         it("Does not create the weightGains and biasGain when the adaptiveLR is not gain", () => {
-            neuron.init(3, "not gain")
+            neuron.init(3, {adaptiveLR: "not gain"})
             expect(neuron.weightGains).to.be.undefined
             expect(neuron.biasGain).to.be.undefined
         })
 
         it("Creates a weightsCache array, with same dimension as weights, if the adaptiveLR is adagrad, with 0 values", () => {
-            neuron.init(3, "adagrad")
+            neuron.init(3, {adaptiveLR: "adagrad"})
             expect(neuron.weightsCache).to.not.be.undefined
             expect(neuron.weightsCache).to.have.lengthOf(3)
             expect(neuron.weightsCache).to.deep.equal([0,0,0])
         })
 
         it("Creates a weightsCache array, with same dimension as weights, if the adaptiveLR is RMSProp, with 0 values", () => {
-            neuron.init(3, "RMSProp")
+            neuron.init(3, {adaptiveLR: "RMSProp"})
             expect(neuron.weightsCache).to.not.be.undefined
             expect(neuron.weightsCache).to.have.lengthOf(3)
             expect(neuron.weightsCache).to.deep.equal([0,0,0])
         })
 
         it("Creates a biasCache value of 0 if the adaptiveLR parameter is adagrad", () => {
-            neuron.init(3, "adagrad")
+            neuron.init(3, {adaptiveLR: "adagrad"})
             expect(neuron.biasCache).to.equal(0)
         })
 
         it("Creates a biasCache value of 0 if the adaptiveLR parameter is RMSProp", () => {
-            neuron.init(3, "adagrad")
+            neuron.init(3, {adaptiveLR: "adagrad"})
             expect(neuron.biasCache).to.equal(0)
         })
 
         it("Does not create the weightsCache or biasCache if the adaptiveLR is not adagrad", () => {
-            neuron.init(3, "not adagrad")
+            neuron.init(3, {adaptiveLR: "not adagrad"})
             expect(neuron.weightsCache).to.be.undefined
             expect(neuron.biasCache).to.be.undefined
         })
 
         it("Does not create the weightsCache or biasCache if the adaptiveLR is not RMSProp", () => {
-            neuron.init(3, "not RMSProp")
+            neuron.init(3, {adaptiveLR: "not RMSProp"})
             expect(neuron.weightsCache).to.be.undefined
             expect(neuron.biasCache).to.be.undefined
         })
 
         it("Creates and sets neuron.m to 0 if the adaptiveLR parameter is adam", () => {
-            neuron.init(3, "adam")
+            neuron.init(3, {adaptiveLR: "adam"})
             expect(neuron.m).to.not.be.undefined
             expect(neuron.m).to.equal(0)
         })
 
         it("Creates and sets neuron.v to 0 if the adaptiveLR parameter is adam", () => {
-            neuron.init(3, "adam")
+            neuron.init(3, {adaptiveLR: "adam"})
             expect(neuron.v).to.not.be.undefined
             expect(neuron.v).to.equal(0)
         })
 
         it("Does not create neuron.m or neuron.v when the adaptiveLR parameter is not adam", () => {
-            neuron.init(3, "not adam")
+            neuron.init(3, {adaptiveLR: "not adam"})
             expect(neuron.m).to.be.undefined
             expect(neuron.v).to.be.undefined
         })
 
         it("Creates a weightsCache array, with same dimension as weights, if the adaptiveLR is adadelta, with 0 values", () => {
-            neuron.init(3, "adadelta")
+            neuron.init(3, {adaptiveLR: "adadelta"})
             expect(neuron.weightsCache).to.not.be.undefined
             expect(neuron.weightsCache).to.have.lengthOf(3)
             expect(neuron.weightsCache).to.deep.equal([0,0,0])
         })
 
         it("Creates a adadeltaBiasCache value of 0 if the adaptiveLR parameter is adadelta", () => {
-            neuron.init(3, "adadelta")
+            neuron.init(3, {adaptiveLR: "adadelta"})
             expect(neuron.adadeltaBiasCache).to.equal(0)
         })
 
         it("Creates a adadeltaCache array, with same dimension as weights, if the adaptiveLR is adadelta, with 0 values", () => {
-            neuron.init(3, "adadelta")
+            neuron.init(3, {adaptiveLR: "adadelta"})
             expect(neuron.adadeltaCache).to.not.be.undefined
             expect(neuron.adadeltaCache).to.have.lengthOf(3)
             expect(neuron.adadeltaCache).to.deep.equal([0,0,0])
         })
 
         it("Does not create adadeltaBiasCache or adadeltaCache when the adaptiveLR is adagrad or RMSProp", () => {
-            neuron.init(3, "adagrad")
+            neuron.init(3, {adaptiveLR: "adagrad"})
             expect(neuron.adadeltaCache).to.be.undefined
             expect(neuron.adadeltaBiasCache).to.be.undefined
             const neuron2 = new Neuron()
-            neuron2.init(3, "RMSProp")
+            neuron2.init(3, {adaptiveLR: "RMSProp"})
             expect(neuron2.adadeltaCache).to.be.undefined
             expect(neuron2.adadeltaBiasCache).to.be.undefined
+        })
+
+        it("Creates a random neuron.rreluSlope number if the activationConfig value is rrelu", () => {
+            neuron.init(3, {activationConfig: "rrelu"})
+            expect(neuron.rreluSlope).to.not.be.undefined
+            expect(neuron.rreluSlope).to.be.a.number
+            expect(neuron.rreluSlope).to.be.at.most(0.0011)
+        })
+
+        it("Sets the neuron.eluAlpha to the given value, if given a value", () => {
+            neuron.init(3, {activationConfig: "elu", eluAlpha: 0.5})
+            expect(neuron.eluAlpha).to.equal(0.5)
         })
     })
 })
@@ -1111,7 +1239,112 @@ describe("Netmath", () => {
         })
         it("sigmoid(0.8430688214048092, true) == 0.21035474941074114", () => {
             expect(NetMath.sigmoid(0.8430688214048092, true)).to.equal(0.21035474941074114)
-        })        
+        })
+    })
+
+    describe("Tanh", () => {
+        it("tanh(1)==0.7615941559557649", () => {
+            expect(NetMath.tanh(1)).to.equal(0.7615941559557649)
+        })
+        it("tanh(0.5)==0.46211715726000974", () => {
+            expect(NetMath.tanh(0.5)).to.equal(0.46211715726000974)
+        })
+        it("tanh(0.5, true)==0.7864477329659275", () => {
+            expect(NetMath.tanh(0.5, true)).to.equal(0.7864477329659275)
+        })
+        it("tanh(1.5, true)==0.18070663892364855", () => {
+            expect(NetMath.tanh(1.5, true)).to.equal(0.18070663892364855)
+        })
+        it("Doesn't return NaN if the input value is too high", () => {
+            expect(NetMath.tanh(900)).to.not.be.NaN
+            expect(NetMath.tanh(900, true)).to.not.be.NaN
+        })
+    })
+
+    describe("relu", () => {
+        it("relu(2)==2", () => {
+            expect(NetMath.relu(2)).to.equal(2)
+        })
+        it("relu(-2)==0", () => {
+            expect(NetMath.relu(-2)).to.equal(0)
+        })
+        it("relu(2, true)==1", () => {
+            expect(NetMath.relu(2, true)).to.equal(1)
+        })
+        it("relu(-2, true)==0", () => {
+            expect(NetMath.relu(-2, true)).to.equal(0)
+        })
+    })
+
+    describe("rrelu", () => {
+        it("rrelu(2, false, {rreluSlope: 0.0005})==2", () => {
+            expect(NetMath.rrelu(2, false, {rreluSlope: 0.0005})).to.equal(2)
+        })
+        it("rrelu(-2, false, {rreluSlope: 0.0005})==0", () => {
+            expect(NetMath.rrelu(-2, false, {rreluSlope: 0.0005})).to.equal(0.0005)
+        })
+        it("rrelu(2, true, {rreluSlope: 0.0005})==1", () => {
+            expect(NetMath.rrelu(2, true, {rreluSlope: 0.0005})).to.equal(1)
+        })
+        it("rrelu(-2, true, {rreluSlope: 0.0005})==0", () => {
+            expect(NetMath.rrelu(-2, true, {rreluSlope: 0.0005})).to.equal(0.0005)
+        })
+    })
+
+    describe("lrelu", () => {
+        it("lrelu(2)==2", () => {
+            expect(NetMath.lrelu.bind({lreluSlope:-0.0005}, 2)()).to.equal(2)
+        })
+        it("lrelu(-2)==-0.001", () => {
+            expect(NetMath.lrelu.bind({lreluSlope:-0.0005}, -2)()).to.equal(-0.001)
+        })
+        it("lrelu(2, true)==1", () => {
+            expect(NetMath.lrelu.bind({lreluSlope:-0.0005}, 2, true)()).to.equal(1)
+        })
+        it("lrelu(-2, true)==0", () => {
+            expect(NetMath.lrelu.bind({lreluSlope:-0.0005}, -2, true)()).to.equal(-0.0005)
+        })
+    })
+
+    describe("sech", () => {
+        it("sech(1)==0.6480542736638853", () => {
+            expect(NetMath.sech(1)).to.equal(0.6480542736638853)
+        })
+        it("sech(-0.5)==0.886818883970074", () => {
+            expect(NetMath.sech(-0.5)).to.equal(0.886818883970074)
+        })
+    })
+
+    describe("lecuntanh", () => {
+        it("lecuntanh(2)==1.4929388053842507", () => {
+            expect(NetMath.lecuntanh(2)).to.equal(1.4929388053842507)
+        })
+        it("lecuntanh(-2)==-1.4929388053842507", () => {
+            expect(NetMath.lecuntanh(-2)).to.equal(-1.4929388053842507)
+        })
+        it("lecuntanh(2, true)==0.2802507761872869", () => {
+            expect(NetMath.lecuntanh(2, true)).to.equal(0.2802507761872869)
+        })
+        it("lecuntanh(-2, true)==0.2802507761872869", () => {
+            expect(NetMath.lecuntanh(-2, true)).to.equal(0.2802507761872869)
+        })
+    })
+
+    describe("elu", () => {
+        it("elu(2)==2", () => {
+            expect(NetMath.elu.bind(null, 2, false, {eluAlpha: 1})()).to.equal(2)
+        })
+
+        it("elu(-0.25)==-0.22119921692859512", () => {
+            expect(NetMath.elu.bind(null, -0.25, false, {eluAlpha: 1})()).to.equal(-0.22119921692859512)
+        })
+
+        it("elu(2, true)==1", () => {
+            expect(NetMath.elu.bind(null, 2, true, {eluAlpha: 1})()).to.equal(1)
+        })
+        it("elu(-0.5, true)==0.6065306597126334", () => {
+            expect(NetMath.elu.bind(null, -0.5, true, {eluAlpha: 1})()).to.equal(0.6065306597126334)
+        })
     })
 
     describe("Cross Entropy", () => {